edited sql scripts pincode

+++ documents/scripts/manages/create/create.sql
@@ -78,7 +78,7 @@ CREATE TABLE IF NOT EXISTS public.addresses (
     city VARCHAR(64),
     state VARCHAR(64),
     country VARCHAR(64),
-    pincode NUMERIC NOT NULL,
+    pincode INTEGER NOT NULL,
     type VARCHAR(64),
     phone VARCHAR(16),
     email VARCHAR(255),
